0-async_generator.py

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -6,7 +6,7 @@ import random
 from typing import Generator
 
 
-async def async_generator() -> Generator[int, None, None]: # type: ignore
+async def async_generator() -> Generator[float, None, None]:  # type: ignore
     """
     loop 10 times, each time asynchronously wait 1 second,
     then yield a random number between 0 and 10
